wb_bidder/requirements.txt
altgraph==0.17.3
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.6
djangorestframework==3.14.0
idna==3.4
macholib==1.16.2
netaddr==0.8.0
pyinstaller-hooks-contrib==2023.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.14.1
yookassa==2.3.4
